Deleting unecessary tests because the class they were testing no llonger exists
